+++ lib/filters.h
@@ -50,7 +50,7 @@ public:
     bool            negated);
   bool match(
     const Node&     node,
-    int             start = 0) const;
+    size_t          start = 0) const;
   /* For verbosity */
   void show(int level = 0) const;
 };
