+++ parsers/cvs_parser.cpp
@@ -86,9 +86,9 @@ CvsParser::CvsParser(Mode mode, const string& dir_path) : Parser(mode, dir_path)
   /* Fill in list of controlled files */
   entries_file.open(O_RDONLY);
   out(debug, msg_standard, "Parsing CVS entries", 1, NULL);
-  int          line_no = 0;
-  char*        buffer = NULL;
-  unsigned int buffer_capacity = 0;
+  int    line_no = 0;
+  char*  buffer = NULL;
+  size_t buffer_capacity = 0;
   while (entries_file.getLine(&buffer, &buffer_capacity) > 0) {
     const char* reader = buffer;
     const char* pos;
